One-Minute Timer
A 1-minute timer that blinks every second for 60 seconds. The timer is a for-loop from 0 (initial blink) to 60 (final blink). Through each loop of the for-loop, the port and pin that is associated with the built-in green LED is toggled, turning on and off. It is followed by a 1000 ms (1 second) delay. Additional toggling and delays after the for-loop are used to confirm that it works as expected (timed and measured to be exactly 1 minute). Wrapped inside an if statement to prepare for further programming features (push inputs to ensure the loop runs only given input from the user).

Next Steps: Configuring the blue pin to be the signal to turn on the device, configuring the black pin to be the number of minutes (additional feature if time), and configuring the 5V pin for output)