Blue Button Toggled for User Input in Minutes
The inputs from the user through the blue button are calibrated and integrated with the timer. The blue push button is used to both set the time in minutes and start the timer. First, the user will press the button for the number of minutes they want to set the timer to. The timer recognizes a press when the button has been pressed and will only recognize the next press if the button is released (holding down the button for 5 seconds will be equivalent to one minute because there has not been a reset)). When they are done setting the timer, they will leave the button alone for 10 seconds (and step away from the device). When at any period, the MCU recognizes that the microcontroller has been idle for 10 seconds, it will start the timer and loop for 60*n seconds where n is the number of minutes that the user inputted. After the timer has expired, the light will stay on and turn off a few seconds after the ejection is complete. The ejection code will be implemented in a later merge.