Ejection Signal is Sent
The Ejection signal is sent through P10. A current is sent to the springs causing them to compress over the span of 5 seconds. The signal is then stopped and this causes the springs to theoretically decompress and push against the wall. This process is then looped 10 times over shorter intervals of 2.5 seconds instead of 5 seconds to make sure the system is ejected. One issue is the signal may not be strong enough to eject our device.